feat: fga-eps-mds/2020.2-Lend.it#127 create send mail
