--- src/config/mail.js
@@ -1,8 +1,0 @@
-const mailTrapConfig = {
-  host: 'smtp.gmail.com',
-  port: 587,
-  user: 'codenomecloves@gmail.com',
-  pass: '27858012',
-};
-
-export default mailTrapConfig;
